+++ chain-abci/src/enclave_bridge/mod.rs
@@ -1,10 +1,9 @@
 use enclave_protocol::{IntraEnclaveRequest, IntraEnclaveResponse};
-use parity_scale_codec::{Decode, Encode};
 
 /// TODO: feature-guard when workspaces can be built with --features flag: https://github.com/rust-lang/cargo/issues/5015
 pub mod mock;
 
-#[cfg(all(not(feature = "mock-validation"), target_os = "linux"))]
+#[cfg(all(not(feature = "mock-enclave"), target_os = "linux"))]
 pub mod real;
 
 /// Abstracts over communication with an external part that does enclave calls